Add initiator type in neo4j. Update usernames db length.
