+++ packages/runner/sync-customer-os-data/entity/conversation.go
@@ -1,0 +1,11 @@
+package entity
+
+type ConversationInitiator struct {
+	Id             string
+	ExternalId     string
+	ExternalSystem string
+	FirstName      string
+	LastName       string
+	Email          string
+	InitiatorType  SenderType
+}
